refactor(parser): extract markdown context scanners
